Backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
dnspython==2.7.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
idna==3.10
motor==3.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
